refactor all the handlers - folder structure
